[JENKINS-62767] Don't include JGit in plugin hpi file
Depend on git plugin 3.3.0 instead of 2.4.1.  The 3.3.0 release is the
last release before the GitSCMSource restrictions that were implemented
in git plugin 3.4.0.

Depend on git client plugin 2.7.0 instead of 1.19.0.  The 2.7.0 release
is an arbitrarily selected release from 2017.  This assumes the currently
installed git client plugin is much newer and that there are no JGit
4.5.4 specific calls in the gitlab-plugin.